requirements.old.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.2
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.1
blis==0.4.1
Bottleneck==1.3.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
efficientnet-pytorch==0.5.1
entrypoints==0.3
fastai==1.0.59
fastprogress==0.1.21
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
idna==2.8
imageio==2.6.1
imgaug==0.3.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.1
numexpr==2.7.0
numpy==1.17.3
nvidia-ml-py3==7.352.0
opencv-python==4.1.1.26
opencv-python-headless==4.1.1.26
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
progressbar2==3.47.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.4.3
pyparsing==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
soupsieve==1.9.5
spacy==2.2.2
srsly==0.2.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.3.1
torch==1.3.0
torchvision==0.4.1
tornado==6.0.3
tqdm==4.38.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
